Revert "YandexMap done"

+++ backend/src/db/migrations/20210917104909-create-location.js
@@ -21,10 +21,10 @@ module.exports = {
         type: Sequelize.TEXT,
       },
       coordX: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.INTEGER,
       },
       coordY: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.INTEGER,
       },
       openingHours: {
         type: Sequelize.TEXT,
